rename eli5.sklearn.unhash to eli5.sklearn.unhashing
This matches sklearn.feature_extraction.hashing name better.
